+++ CovidStateDashboard/snowflakeQuery/index.js
